Merapikan letak kode
Menambah cara compile
