+++ README.md
@@ -1,2 +1,11 @@
 # Tubes-Alstrukdat
 Engi's Kitchen DLC
+
+jika berada di folder root / luar src compile dengan
+'''
+gcc src/*.c -o bin/game
+'''
+atau jika berada didalam folder bin
+'''
+gcc ../src/*.c -o game
+'''
