+++ src/main.c
@@ -1,0 +1,7 @@
+#include<stdio.h>
+
+int main()
+{
+
+    return 0;
+}
