+++ src/stackt.c
@@ -1,16 +1,10 @@
-/* 	NIM/Nama : 13517040/Ariel Ansa Razumardi
-	Nama file : stackt.c
-	Topik : Pra Praktikum 07
-	Tanggal : 5 Oktober 2018
-	Deskripsi : Implementasi semua fungsi dan prosedur yang sudah didefinisikan di file stackt.h */
-
-#include "stackt.h"
+#include "../include/stackt.h"
 #include <stdio.h>
 
 
 /* ************ Prototype ************ */
 /* *** Konstruktor/Kreator *** */
-void CreateEmpty (Stack *S){
+void CreateEmptyStack (Stack *S){
     Top(*S) = Nil;
 }
 /* I.S. sembarang; */
@@ -19,11 +13,11 @@ void CreateEmpty (Stack *S){
 /* Ciri stack kosong : TOP bernilai Nil */
 
 /* ************ Predikat Untuk test keadaan KOLEKSI ************ */
-boolean IsEmpty (Stack S){
+boolean IsEmptyStack (Stack S){
     return (Top(S) == Nil);
 }
 /* Mengirim true jika Stack kosong: lihat definisi di atas */
-boolean IsFull (Stack S){
+boolean IsFullStack (Stack S){
     return (Top(S) == MaxEl); 
 }
 /* Mengirim true jika tabel penampung nilai elemen stack penuh */
